* Form Hacker did not use FORMID so when you copied a form, it would break * Updated widget classes (may "break" existing forms) * Added a "link" to the Auto Respond Email widget that allows users to set a custom field on a post. Can be used to verify if the user email is valid. * Fixed critical Windows host bug that would attempt to delete root drive. The add_post_meta Wordpress function would strip back slashes out of input and basically feck up the Windows path. Now the path name is "protected" before being passed to add_post_meta. *Fixed post queuing. This was broken in two ways. The date/time calculation was wrong and now has been updated based on generousily donated code from [Adam Selvidge]( http://www.myconfinedspace.com/ ). Second a change in Wordpress 2.7 meant that setting the future status was being ignored when the post was being published.
git-svn-id: https://plugins.svn.wordpress.org/tdo-mini-forms/trunk@94302 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ include/tdomf-notify.php
@@ -218,6 +218,11 @@ function tdomf_notify_admins($post_ID,$form_id){
 function tdomf_notify_poster_approved($post_id) {
    global $wpdb;
    
+   if(wp_is_post_revision($post_id)) {
+       tdomf_log_message_extra("tdomf_notify_poster_approved: post $post_id is a revision -- do nothing.");
+       return $post_id;
+   }
+   
    $email = get_post_meta($post_id, TDOMF_KEY_NOTIFY_EMAIL, true);
    delete_post_meta($post_id, TDOMF_KEY_NOTIFY_EMAIL);
 
@@ -265,6 +270,11 @@ function tdomf_notify_poster_approved($post_id) {
 //
 function tdomf_notify_poster_rejected($post_id) {
    global $wpdb;
+   
+   if(wp_is_post_revision($post_id)) {
+       tdomf_log_message_extra("tdomf_notify_poster_rejected: post $post_id is a revision -- do nothing.");
+       return $post_id;
+   }
    
    tdomf_log_message_extra("tdomf_notify_poster_rejected: $email");
